+++ src/main/java/com/fourweekdays/fourweekdays/inventory/controller/InventoryController.java
@@ -2,22 +2,26 @@ package com.fourweekdays.fourweekdays.inventory.controller;
 
 import com.fourweekdays.fourweekdays.common.BaseResponse;
 import com.fourweekdays.fourweekdays.inventory.model.dto.request.InventorySearchRequest;
-import com.fourweekdays.fourweekdays.inventory.model.dto.response.InventoryListDto;
 import com.fourweekdays.fourweekdays.inventory.model.dto.response.InventoryReadDto;
+import com.fourweekdays.fourweekdays.inventory.service.InventoryService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
+@RequestMapping("/api/inventories")
+@RequiredArgsConstructor
 @RestController
-@RequestMapping("/api/inventories") // 복수형으로 작성
 public class InventoryController {
 
-    // 재고 목록 조회
+    private final InventoryService inventoryService;
+
     @GetMapping
-    public ResponseEntity<BaseResponse<InventoryListDto>> inventoryList(InventorySearchRequest request) {
-        return ResponseEntity.ok(BaseResponse.success(new InventoryListDto()));
+    public ResponseEntity<BaseResponse<Page<InventoryReadDto>>> inventoryList(@RequestParam(defaultValue = "0") int page,
+                                                            @RequestParam(defaultValue = "10") int size,
+                                                            @RequestBody InventorySearchRequest request) {
+        Page<InventoryReadDto> inventories = inventoryService.searchInventory(request, page, size);
+        return ResponseEntity.ok(BaseResponse.success(inventories));
     }
 
     // 재고 상세 조회
